requirements.txt
fastapi
uvicorn[standard]
streamlit>=1.28.0
python-docx>=1.1.0
openai
python-dotenv>=1.0.0
requests
aiofiles
litellm>=1.0.0
python-multipart
plotly>=5.17.0
pandas>=2.0.0

# Security dependencies
defusedxml>=0.7.1
python-magic>=0.4.27
slowapi>=0.1.9
jsonschema>=4.20.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0
httpx>=0.24.0
faker>=19.0.0